add quantity and maxQuantity at CartItem

+++ app/src/main/java/com/vishalgaur/shoppingapp/ui/home/CartItemAdapter.kt
@@ -49,7 +49,7 @@ class CartItemAdapter(
 			}
 			binding.cartProductMinusBtn.setOnClickListener {
 				binding.loaderLayout.loaderFrameLayout.visibility = View.VISIBLE
-				onClickListener.onMinusClick(itemData.itemId, itemData.quantity, binding.loaderLayout)
+				onClickListener.onMinusClick(itemData.itemId, itemData.quantity.toInt(), binding.loaderLayout)
 			}
 
 		}
