+++ app/src/main/java/com/vishalgaur/shoppingapp/ui/home/CartItemAdapterWithLike.kt
@@ -66,7 +66,7 @@ class CartItemAdapterWithLike(
 			}
 			binding.cartProductMinusBtn.setOnClickListener {
 				binding.loaderLayout.loaderFrameLayout.visibility = View.VISIBLE
-				onClickListener.onMinusClick(itemData.itemId, itemData.quantity, binding.loaderLayout)
+				onClickListener.onMinusClick(itemData.itemId, itemData.quantity.toInt(), binding.loaderLayout)
 			}
 
 		}
